Extract collaboration routes into dedicated module
