+++ TASKS.md
@@ -520,7 +520,7 @@ The repository has grown organically and requires systematic refactoring to impr
   - [ ] Extract project management routes to `routes/projects.py` (list, get, create, update, delete, collaborators, templates, export)
   - [x] Extract marketplace routes to `routes/marketplace.py` (list, get, publish, reviews, search)
   - [x] Extract forum routes to `routes/forum.py` (list threads, create thread, create post, search)
-  - [ ] Extract collaboration routes to `routes/collaboration.py` (sessions, presence, heartbeat, comments)
+  - [x] Extract collaboration routes to `routes/collaboration.py` (sessions, presence, heartbeat, comments)
   - [x] Extract asset management routes to `routes/assets.py` (list, upload, download, delete)
   - [x] Extract playtest/testing routes to `routes/playtest.py` (websocket, transcript, replay)
 - [x] Extract health check routes to `routes/health.py` (health, readiness endpoints)
