+++ src/textadventure/api/app.py
@@ -64,6 +64,7 @@ from .backup import BackupUploadMetadata, BackupUploader, S3BackupUploader
 from .settings import SceneApiSettings
 from .routes import (
     create_assets_router,
+    create_collaboration_router,
     create_playtest_router,
     create_health_router,
     create_marketplace_router,
@@ -134,9 +135,6 @@ from .models import (
     SceneCommentResource,
     SceneCommentThreadResource,
     SceneCommentThreadListResponse,
-    SceneCommentThreadCreateRequest,
-    SceneCommentReplyRequest,
-    SceneCommentResolveRequest,
     # Project models
     AdventureProjectResource,
     AdventureProjectListResponse,
@@ -151,7 +149,6 @@ from .models import (
     ProjectCollaboratorUpdateRequest,
     ProjectCollaborationSessionResource,
     ProjectCollaborationSessionListResponse,
-    ProjectCollaborationSessionRequest,
     # Marketplace models
     MarketplaceEntrySummary,
     MarketplaceReview,
@@ -6090,6 +6087,13 @@ def create_app(
         )
     )
 
+    app.include_router(  # type: ignore[attr-defined]
+        create_collaboration_router(
+            project_service=project,
+            comment_service=comment,
+        )
+    )
+
     @app.get(
         "/api/scenes",
         response_model=SceneListResponse,
@@ -6756,246 +6760,6 @@ def create_app(
             raise HTTPException(status_code=403, detail=str(exc)) from exc
         except RuntimeError as exc:
             raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    @app.get(
-        "/api/projects/{project_id}/collaboration/sessions",
-        response_model=ProjectCollaborationSessionListResponse,
-        tags=["Projects"],
-    )
-    def list_project_collaboration_sessions(
-        project_id: str,
-    ) -> ProjectCollaborationSessionListResponse:
-        if project is None:
-            raise HTTPException(404, "Project management endpoints are not enabled.")
-
-        try:
-            return project.list_collaboration_sessions(project_id)
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-        except RuntimeError as exc:
-            raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    @app.post(
-        "/api/projects/{project_id}/collaboration/sessions",
-        response_model=ProjectCollaborationSessionListResponse,
-        tags=["Projects"],
-    )
-    def touch_project_collaboration_session(
-        project_id: str,
-        payload: ProjectCollaborationSessionRequest,
-        acting_user_id: str | None = Query(
-            None,
-            description=(
-                "Identifier of the collaborator performing the join or heartbeat."
-            ),
-        ),
-    ) -> ProjectCollaborationSessionListResponse:
-        if project is None:
-            raise HTTPException(404, "Project management endpoints are not enabled.")
-
-        try:
-            return project.touch_collaboration_session(
-                project_id,
-                acting_user_id=acting_user_id,
-                session_id=payload.session_id,
-                scene_id=payload.scene_id,
-                ttl_seconds=payload.ttl_seconds,
-            )
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-        except ProjectPermissionError as exc:
-            raise HTTPException(status_code=403, detail=str(exc)) from exc
-        except RuntimeError as exc:
-            raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    @app.delete(
-        "/api/projects/{project_id}/collaboration/sessions/{session_id}",
-        response_model=ProjectCollaborationSessionListResponse,
-        tags=["Projects"],
-    )
-    def delete_project_collaboration_session(
-        project_id: str,
-        session_id: str,
-        acting_user_id: str | None = Query(
-            None,
-            description=(
-                "Identifier of the collaborator performing the session termination."
-            ),
-        ),
-    ) -> ProjectCollaborationSessionListResponse:
-        if project is None:
-            raise HTTPException(404, "Project management endpoints are not enabled.")
-
-        try:
-            return project.end_collaboration_session(
-                project_id,
-                session_id,
-                acting_user_id=acting_user_id,
-            )
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-        except ProjectPermissionError as exc:
-            raise HTTPException(status_code=403, detail=str(exc)) from exc
-        except RuntimeError as exc:
-            raise HTTPException(status_code=500, detail=str(exc)) from exc
-
-    @app.get(
-        "/api/projects/{project_id}/scenes/{scene_id}/comments",
-        response_model=SceneCommentThreadListResponse,
-        tags=["Scene Comments"],
-    )
-    def list_scene_comment_threads(
-        project_id: str,
-        scene_id: str,
-        location_type: SceneCommentLocationType | None = Query(
-            None,
-            description="Optional location type filter when listing comment threads.",
-        ),
-        choice_command: str | None = Query(
-            None,
-            description="Optional transition command filter for comment threads.",
-        ),
-    ) -> SceneCommentThreadListResponse:
-        if comment is None:
-            raise HTTPException(
-                status_code=404,
-                detail="Scene comment functionality is not configured for this deployment.",
-            )
-
-        try:
-            return comment.list_threads(
-                project_id,
-                scene_id,
-                location_type=location_type,
-                choice_command=choice_command,
-            )
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    @app.post(
-        "/api/projects/{project_id}/scenes/{scene_id}/comments",
-        response_model=SceneCommentThreadResource,
-        status_code=201,
-        tags=["Scene Comments"],
-    )
-    def create_scene_comment_thread(
-        project_id: str,
-        scene_id: str,
-        payload: SceneCommentThreadCreateRequest,
-        acting_user_id: str | None = Query(
-            None,
-            description="Identifier of the collaborator creating the comment thread.",
-        ),
-    ) -> SceneCommentThreadResource:
-        if comment is None:
-            raise HTTPException(
-                status_code=404,
-                detail="Scene comment functionality is not configured for this deployment.",
-            )
-
-        try:
-            return comment.create_thread(
-                project_id,
-                scene_id,
-                location=payload.location,
-                body=payload.body,
-                author_id=payload.author_id or acting_user_id,
-                author_display_name=payload.author_display_name,
-                acting_user_id=acting_user_id,
-            )
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ProjectPermissionError as exc:
-            raise HTTPException(status_code=403, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    @app.post(
-        "/api/projects/{project_id}/scenes/{scene_id}/comments/{thread_id}/replies",
-        response_model=SceneCommentThreadResource,
-        status_code=201,
-        tags=["Scene Comments"],
-    )
-    def add_scene_comment_reply(
-        project_id: str,
-        scene_id: str,
-        thread_id: str,
-        payload: SceneCommentReplyRequest,
-        acting_user_id: str | None = Query(
-            None,
-            description="Identifier of the collaborator adding the inline comment reply.",
-        ),
-    ) -> SceneCommentThreadResource:
-        if comment is None:
-            raise HTTPException(
-                status_code=404,
-                detail="Scene comment functionality is not configured for this deployment.",
-            )
-
-        try:
-            return comment.add_comment(
-                project_id,
-                scene_id,
-                thread_id,
-                body=payload.body,
-                author_id=payload.author_id or acting_user_id,
-                author_display_name=payload.author_display_name,
-                acting_user_id=acting_user_id,
-            )
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except KeyError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ProjectPermissionError as exc:
-            raise HTTPException(status_code=403, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    @app.post(
-        "/api/projects/{project_id}/scenes/{scene_id}/comments/{thread_id}/resolution",
-        response_model=SceneCommentThreadResource,
-        tags=["Scene Comments"],
-    )
-    def set_scene_comment_resolution(
-        project_id: str,
-        scene_id: str,
-        thread_id: str,
-        payload: SceneCommentResolveRequest,
-        acting_user_id: str | None = Query(
-            None,
-            description="Identifier of the collaborator updating the comment thread resolution state.",
-        ),
-    ) -> SceneCommentThreadResource:
-        if comment is None:
-            raise HTTPException(
-                status_code=404,
-                detail="Scene comment functionality is not configured for this deployment.",
-            )
-
-        try:
-            return comment.set_resolution(
-                project_id,
-                scene_id,
-                thread_id,
-                resolved=payload.resolved,
-                acting_user_id=acting_user_id,
-            )
-        except FileNotFoundError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except KeyError as exc:
-            raise HTTPException(status_code=404, detail=str(exc)) from exc
-        except ProjectPermissionError as exc:
-            raise HTTPException(status_code=403, detail=str(exc)) from exc
-        except ValueError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
 
     app.include_router(  # type: ignore[attr-defined]
         create_marketplace_router(
