+++ src/textadventure/api/routes/__init__.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from .assets import create_assets_router
+from .collaboration import create_collaboration_router
 from .forum import create_forum_router
 from .health import create_health_router
 from .marketplace import create_marketplace_router
@@ -13,6 +14,7 @@ from .users import create_users_router
 
 __all__ = [
     "create_assets_router",
+    "create_collaboration_router",
     "create_forum_router",
     "create_health_router",
     "create_marketplace_router",
